[Learn][-g, --max-leaf-gamma] limit the maximum gamma in a leaf
This option greatly enhances the stability of the learning process,
primarily in the context of logistic models, which are prone to
diverging where there are regions of the feature space where the
probability value approaches either 0 or 1, and where correspondingly
the gradient diverges. This works by constraining the gamma values
produced by the splitter into a reasonable range. While this might
reduce the speed of convergence in some cases, an appropriate choice of
max gamma generally enhances model stability.